practice_final/03_04/requirements.txt
contourpy==1.0.7
cycler==0.11.0
et-xmlfile==1.1.0
fonttools==4.38.0
kiwisolver==1.4.4
matplotlib==3.6.3
numpy==1.24.1
openpyxl==3.1.2
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
seaborn==0.12.2
six==1.16.0
tomli==2.0.1
